Support specifying local IPv6 adddress/interface
...for outgoing packets.

This is the first step in supporting IPV6_PKTINFO in the send path.

Note that the implementation enforces strong host semantics as gVisor
currently follows the strong host model.

This change does not support IPV6_PKTINFO in the send path from Linux
applications (via CMSGS) yet as there is no need. The bindings to enable
the IPV6_PKTINFO CMSG in the sendpath can be done in a followup.

Bug: https://fxbug.dev/102222
PiperOrigin-RevId: 454018146